+++ ompi/mca/fs/base/base.h
@@ -10,7 +10,7 @@
  * Copyright (c) 2004-2005 The Regents of the University of California.
  *                         All rights reserved.
  * Copyright (c) 2008      Cisco Systems, Inc.  All rights reserved.
- * Copyright (c) 2008-2011 University of Houston. All rights reserved.
+ * Copyright (c) 2008-2018 University of Houston. All rights reserved.
  * Copyright (c) 2018      Research Organization for Information Science
  *                         and Technology (RIST). All rights reserved.
  * $COPYRIGHT$
@@ -52,6 +52,13 @@ OMPI_DECLSPEC int mca_fs_base_init_file (struct ompio_file_t *file);
 OMPI_DECLSPEC int mca_fs_base_get_param (struct ompio_file_t *file, int keyval);
 OMPI_DECLSPEC void mca_fs_base_get_parent_dir (char *filename, char **dirnamep);
 OMPI_DECLSPEC int  mca_fs_base_get_fstype(char *fname);
+
+OMPI_DECLSPEC int mca_fs_base_file_delete (char* file_name, struct opal_info_t *info);
+OMPI_DECLSPEC int mca_fs_base_file_sync (ompio_file_t *fh);
+OMPI_DECLSPEC int mca_fs_base_file_get_size (ompio_file_t *fh, OMPI_MPI_OFFSET_TYPE *size);
+OMPI_DECLSPEC int mca_fs_base_file_set_size (ompio_file_t *fh, OMPI_MPI_OFFSET_TYPE size);
+OMPI_DECLSPEC int mca_fs_base_file_close (ompio_file_t *fh);
+
 /*
  * Globals
  */
