+++ ompi/mca/fs/base/fs_base_file_delete.c
@@ -2,16 +2,15 @@
  * Copyright (c) 2004-2005 The Trustees of Indiana University and Indiana
  *                         University Research and Technology
  *                         Corporation.  All rights reserved.
- * Copyright (c) 2004-2017 The University of Tennessee and The University
+ * Copyright (c) 2004-2011 The University of Tennessee and The University
  *                         of Tennessee Research Foundation.  All rights
  *                         reserved.
  * Copyright (c) 2004-2005 High Performance Computing Center Stuttgart,
  *                         University of Stuttgart.  All rights reserved.
  * Copyright (c) 2004-2005 The Regents of the University of California.
  *                         All rights reserved.
- * Copyright (c) 2008-2011 University of Houston. All rights reserved.
- * Copyright (c) 2018      Research Organization for Information Science
- *                         and Technology (RIST). All rights reserved.
+ * Copyright (c) 2008-2018 University of Houston. All rights reserved.
+ * Copyright (c) 2016-2017 IBM Corporation. All rights reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -21,32 +20,30 @@
 
 
 #include "ompi_config.h"
-#include "fs_ufs.h"
+#include "base.h"
 
-#include <fcntl.h>
 #include <unistd.h>
+
 #include "mpi.h"
 #include "ompi/constants.h"
 #include "ompi/mca/fs/fs.h"
 
 /*
- *	file_close_ufs
+ *	file_delete_ufs
  *
- *	Function:	- closes a new file
- *	Accepts:	- file handle
+ *	Function:	- deletes a file
+ *	Accepts:	- file name & info
  *	Returns:	- Success if file closed
  */
-int
-mca_fs_ufs_file_close (ompio_file_t *fh)
+int mca_fs_base_file_delete (char* file_name,
+                             struct opal_info_t *info)
 {
-    fh->f_comm->c_coll->coll_barrier (fh->f_comm,
-                                     fh->f_comm->c_coll->coll_barrier_module);
-    /*    close (*(int *)fh->fd);*/
-    close (fh->fd);
-    /*    if (NULL != fh->fd)
-    {
-        free (fh->fd);
-        fh->fd = NULL;
-        }*/
+    int ret;
+
+    ret = unlink(file_name);
+    if (0 > ret) {
+        return OMPI_ERROR;
+    }
+
     return OMPI_SUCCESS;
 }
