+++ ompi/mca/fs/base/fs_base_file_sync.c
@@ -2,14 +2,14 @@
  * Copyright (c) 2004-2005 The Trustees of Indiana University and Indiana
  *                         University Research and Technology
  *                         Corporation.  All rights reserved.
- * Copyright (c) 2004-2017 The University of Tennessee and The University
+ * Copyright (c) 2004-2011 The University of Tennessee and The University
  *                         of Tennessee Research Foundation.  All rights
  *                         reserved.
  * Copyright (c) 2004-2005 High Performance Computing Center Stuttgart,
  *                         University of Stuttgart.  All rights reserved.
  * Copyright (c) 2004-2005 The Regents of the University of California.
  *                         All rights reserved.
- * Copyright (c) 2008-2011 University of Houston. All rights reserved.
+ * Copyright (c) 2008-2018 University of Houston. All rights reserved.
  * Copyright (c) 2018      Research Organization for Information Science
  *                         and Technology (RIST). All rights reserved.
  * $COPYRIGHT$
@@ -21,33 +21,20 @@
 
 
 #include "ompi_config.h"
-#include "fs_lustre.h"
+#include "base.h"
+
+#include <unistd.h>
 
 #include "mpi.h"
 #include "ompi/constants.h"
 #include "ompi/mca/fs/fs.h"
 
-/*
- *	file_set_size_lustre
- *
- *	Function:	- set_size of a file
- *	Accepts:	- same arguments as MPI_File_set_size()
- *	Returns:	- Success if size is set
- */
-int
-mca_fs_lustre_file_set_size (ompio_file_t *fh,
-                         OMPI_MPI_OFFSET_TYPE size)
+int mca_fs_base_file_sync (ompio_file_t *fh)
 {
-    int err = 0;
+    int err;
 
-    err = ftruncate(fh->fd, size);
+    err = fsync(fh->fd);
 
-    fh->f_comm->c_coll->coll_bcast (&err,
-                                   1,
-                                   MPI_INT,
-                                   OMPIO_ROOT,
-                                   fh->f_comm,
-                                   fh->f_comm->c_coll->coll_bcast_module);
     if (-1 == err) {
         return OMPI_ERROR;
     }
