--- ompi/mca/fs/ufs/fs_ufs_file_sync.c
@@ -1,42 +1,0 @@
-/*
- * Copyright (c) 2004-2005 The Trustees of Indiana University and Indiana
- *                         University Research and Technology
- *                         Corporation.  All rights reserved.
- * Copyright (c) 2004-2011 The University of Tennessee and The University
- *                         of Tennessee Research Foundation.  All rights
- *                         reserved.
- * Copyright (c) 2004-2005 High Performance Computing Center Stuttgart,
- *                         University of Stuttgart.  All rights reserved.
- * Copyright (c) 2004-2005 The Regents of the University of California.
- *                         All rights reserved.
- * Copyright (c) 2008-2011 University of Houston. All rights reserved.
- * Copyright (c) 2018      Research Organization for Information Science
- *                         and Technology (RIST). All rights reserved.
- * $COPYRIGHT$
- *
- * Additional copyrights may follow
- *
- * $HEADER$
- */
-
-
-#include "ompi_config.h"
-#include "fs_ufs.h"
-#include <unistd.h>
-
-#include "mpi.h"
-#include "ompi/constants.h"
-#include "ompi/mca/fs/fs.h"
-
-int
-mca_fs_ufs_file_sync (ompio_file_t *fh)
-{
-    int err;
-
-    err = fsync(fh->fd);
-
-    if (-1 == err) {
-        return OMPI_ERROR;
-    }
-    return OMPI_SUCCESS;
-}
